stream-processing/requirements.txt
pyspark>=3.5.0
redis>=5.0.0
psycopg2-binary>=2.9.0
kafka-python>=2.0.0
